@@ -1,41 +1,42 @@
 package com.zuhlke.command;
 
+import com.zuhlke.exception.InvalidInputException;
 import com.zuhlke.model.Canvas;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
 import java.io.PrintStream;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class DrawLineTest {
+class CreateCanvasCommandTest {
 
     private Canvas canvas;
-    private DrawLine subject;
+    private CreateCanvasCommand subject;
     private String br;
     private String ans;
 
-    @Before
-    public void setUp() {
+    @BeforeEach
+    void setUp() {
         canvas = new Canvas(20, 4);
-        subject = new DrawLine();
+        subject = new CreateCanvasCommand();
         br = System.getProperty("line.separator");
         ans = br +
                 "----------------------" + br +
                 "|                    |" + br +
-                "|xxxxxx              |" + br +
+                "|                    |" + br +
                 "|                    |" + br +
                 "|                    |" + br +
                 "----------------------" + br;
-
     }
 
     @Test
-    public void drawLineImageTest() {
-        String command = "L 1 2 6 2";
+    void setCanvasImageTest() {
+        String command = "C 20 4";
 
         //prepare to redirect output
         OutputStream os = new ByteArrayOutputStream();
@@ -43,7 +44,8 @@ public class DrawLineTest {
         System.setOut(ps);
 
         try {
-            canvas = subject.Execute(command.split(" "), canvas);
+            canvas = subject.execute(command.split(" "), canvas);
+            canvas.print();
         } catch (Exception e) {
             System.out.println(e.getMessage());
         }
@@ -52,26 +54,15 @@ public class DrawLineTest {
     }
 
     @Test
-    public void drawLineInReverseImageTest() {
-        Canvas testCanvas = new Canvas(20, 4);
-        String command = "L 6 2 1 2";
-
-        //prepare to redirect output
-        OutputStream os = new ByteArrayOutputStream();
-        PrintStream ps = new PrintStream(os);
-        System.setOut(ps);
-
-        try {
-            canvas = subject.Execute(command.split(" "), testCanvas);
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
-
-        assertEquals(ans, os.toString());
+    void checkValidCanvasParameters() {
+        Assertions.assertThrows(InvalidInputException.class, () -> {
+            String command = "C 20 -5";
+            subject.execute(command.split(" "), canvas);
+        });
     }
 
-    @After
-    public void tearDown() {
+    @AfterEach
+    void tearDown() {
         //Restore normal output
         System.setOut(System.out);
     }
