@@ -1,52 +1,47 @@
 package com.zuhlke.model;
 
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
 import java.io.PrintStream;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class CoordinateTest {
+class CoordinateTest {
 
     private Coordinate start;
     private Coordinate end;
 
-    @Before
-    public void setUp() {
+    @BeforeEach
+    void setUp() {
         start = new Coordinate(2, 3);
         end = new Coordinate(5, 12);
     }
 
     @Test
-    public void ExpectGetDistanceOf3and9() {
+    void ExpectGetDistanceOf3and9() {
         Coordinate answer = new Coordinate(5 - 2, 12 - 3);
         assertEquals(answer.getX(), end.getDistance(start).getX());
         assertEquals(answer.getY(), end.getDistance(start).getY());
     }
 
     @Test
-    public void ExpectStartLinearDistance5() {
+    void ExpectStartLinearDistance5() {
         int ans = start.getX() + start.getY();
         assertEquals(ans, start.linearDistance());
     }
 
     @Test
-    public void ExpectLinearDistance10() {
+    void ExpectLinearDistance10() {
         Coordinate point = new Coordinate(3, 7);
         assertEquals(10, point.linearDistance());
     }
 
-    @Test
-    public void ExpectTestEqualsStart() {
-        Coordinate test = new Coordinate(2, 3);
-        assertTrue(test.equals(start));
-    }
 
     @Test
-    public void ExpectPrintOutOfCoordinates() {
+    void ExpectPrintOutOfCoordinates() {
 
         String ans = "X-Coordinate: 2, Y-Coordinate: 3" + System.getProperty("line.separator");
 
@@ -55,15 +50,15 @@ public class CoordinateTest {
         PrintStream ps = new PrintStream(os);
         System.setOut(ps);
 
-        start.print();
+        System.out.println(start.toString());
 
         assertEquals(ans, os.toString());
 
     }
 
     @Test
-    public void addOneToX() {
-        Coordinate test = new Coordinate(3,4);
+    void addOneToX() {
+        Coordinate test = new Coordinate(3, 4);
 
         assertEquals(4, test.addX(1));
         assertEquals(3, test.addX(-1));
@@ -71,8 +66,8 @@ public class CoordinateTest {
     }
 
     @Test
-    public void addOneToY() {
-        Coordinate test = new Coordinate(3,4);
+    void addOneToY() {
+        Coordinate test = new Coordinate(3, 4);
         assertEquals(5, test.addY(1));
         assertEquals(4, test.addY(-1));
     }
