@@ -3,6 +3,6 @@ package com.zuhlke.command;
 import com.zuhlke.model.Canvas;
 
 public interface Command {
-    Canvas Execute(String[] input, Canvas base)throws Exception;
+    Canvas execute(String[] input, Canvas source) throws Exception;
 
 }
