@@ -6,7 +6,21 @@ public class Canvas {
     private int row;
     private int column;
 
-    public Canvas() {
+    //copy constructor
+    public Canvas(Canvas canvas) {
+        this.row = canvas.getRow();
+        this.column = canvas.getColumn();
+        this.base = new Character[row][column];
+
+        for (int i = 0; i < row; i++) {
+            for (int j = 0; j < column; j++) {
+                base[i][j] = canvas.getBase()[i][j];
+            }
+        }
+    }
+
+    private int getRow() {
+        return row;
     }
 
     public Canvas(int x, int y) {
@@ -16,16 +30,12 @@ public class Canvas {
         makeOutline();
     }
 
-    public Character[][] getBase() {
-        return base;
-    }
-
-    public int getRow() {
-        return row;
-    }
-
-    public int getColumn() {
+    private int getColumn() {
         return column;
+    }
+
+    Character[][] getBase() {
+        return base;
     }
 
     //draw outline of the canvas
@@ -52,21 +62,19 @@ public class Canvas {
     }
 
     //plot a single point in the canvas
-    public Character[][] plot(Coordinate point, Character x) {
+    public void plot(Coordinate point, Character x) {
         base[point.getY()][point.getX()] = x;
-        return getBase();
     }
 
     //retrieve character at a single point in the canvas
     public Character getCell(Coordinate point) {
-
         return base[point.getY()][point.getX()];
 
     }
 
     //check if a coordinate lies within canvas
     public boolean isValidPoint(Coordinate point) {
-        return (point.getX() >= 1 && point.getX() <= column && point.getY() >= 1 && point.getY() <= row);
+        return (point.getX() >= 1 && point.getX() < column - 1 && point.getY() >= 1 && point.getY() < row - 1);
     }
 
     public void print() {
@@ -78,5 +86,33 @@ public class Canvas {
         }
         System.out.println();
 
+    }
+
+    public Canvas drawLine(Coordinate start, Coordinate end) {
+        Coordinate diff = end.getDistance(start);
+
+        //loop through half of the linear distance
+        for (int i = 0; i <= diff.linearDistance() / 2; i++) {
+            this.plot(start, 'x');
+            this.plot(end, 'x');
+
+            //if else statements to move coordinates of end and start closer to each other
+            if (diff.getX() > 0) {
+                start.addX(1);
+                end.addX(-1);
+            } else if (diff.getX() < 0) {
+                end.addX(1);
+                start.addX(-1);
+            }
+
+            if (diff.getY() > 0) {
+                start.addY(1);
+                end.addY(-1);
+            } else if (diff.getY() < 0) {
+                end.addY(1);
+                start.addY(-1);
+            }
+        }
+        return this;
     }
 }
