@@ -1,6 +1,6 @@
 package com.zuhlke.model;
 
-public class Coordinate {
+public final class Coordinate {
 
     private int x;
     private int y;
@@ -8,7 +8,6 @@ public class Coordinate {
     public int getX() {
         return x;
     }
-
 
     public int getY() {
         return y;
@@ -20,11 +19,6 @@ public class Coordinate {
 
     public int addY(int i) {
         return this.y = this.y + i;
-    }
-
-    public Coordinate() {
-        this.x = 0;
-        this.y = 0;
     }
 
     public Coordinate(int x, int y) {
@@ -41,13 +35,9 @@ public class Coordinate {
     }
 
     @Override
-    public boolean equals(Object obj) {
-        Coordinate ref = (Coordinate) obj;
-        return (x == ref.x && y == ref.y);
+    public String toString() {
+        return String.format("X-Coordinate: %s, Y-Coordinate: %s", this.x, this.y);
     }
 
-    public void print() {
-        System.out.println(String.format("X-Coordinate: %s, Y-Coordinate: %s", this.x, this.y));
-    }
 
 }
